+++ ErsatzTV.Application/Streaming/HlsSessionWorker.cs
@@ -54,6 +54,8 @@ public class HlsSessionWorker : IHlsSessionWorker
     {
         lock (_sync)
         {
+            _logger.LogDebug("Keep alive - session worker for channel {ChannelNumber}", _channelNumber);
+            
             _lastAccess = DateTimeOffset.Now;
 
             _timer?.Stop();